separates NetworkError into its own file, rewrites getSummaryFor to use Result
