+++ Daydream/Networking/NetworkError.swift
@@ -1,0 +1,18 @@
+//
+//  NetworkError.swift
+//  Daydream
+//
+//  Created by Raymond Kim on 6/4/20.
+//  Copyright © 2020 Raymond Kim. All rights reserved.
+//
+
+import Foundation
+
+enum NetworkError: Error {
+    case malformedJSON
+    case insufficientResults
+    case malformedPhotoField
+    case photoMetadataMissing
+    case APIKeysFetchFailure
+    case unknown // e.g., 3rd party function returns nil data and nil error
+}
